fix: remove h2 dependency
Fixes gh-4802
