+++ spring-cloud-dataflow-server-core/pom.xml
@@ -141,6 +141,7 @@
 		<dependency>
 			<groupId>com.h2database</groupId>
 			<artifactId>h2</artifactId>
+			<optional>true</optional>
 		</dependency>
 		<dependency>
 			<groupId>com.oracle.ojdbc</groupId>
