+++ spring-cloud-dataflow-server/pom.xml
@@ -118,6 +118,11 @@
 			<artifactId>mssqlserver</artifactId>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+			<groupId>com.h2database</groupId>
+			<artifactId>h2</artifactId>
+			<scope>test</scope>
+		</dependency>
 	</dependencies>
 
 	<build>
