+++ spring-cloud-dataflow-shell-core/pom.xml
@@ -91,5 +91,10 @@
 			<version>1.1.2</version>
 			<scope>test</scope>
 		</dependency>
+		<dependency>
+			<groupId>com.h2database</groupId>
+			<artifactId>h2</artifactId>
+			<scope>test</scope>
+		</dependency>
 	</dependencies>
 </project>
